+++ crates/miden-lib/src/testing/mod.rs
@@ -1,4 +1,5 @@
 pub mod account_component;
 pub mod mock_account;
 pub mod mock_account_code;
+pub mod mock_util_lib;
 pub mod note;