+++ crates/miden-objects/src/asset/mod.rs
@@ -15,13 +15,14 @@ use alloc::boxed::Box;
 pub use fungible::FungibleAsset;
 
 mod nonfungible;
+
 pub use nonfungible::{NonFungibleAsset, NonFungibleAssetDetails};
 
 mod token_symbol;
 pub use token_symbol::TokenSymbol;
 
 mod vault;
-pub use vault::{AssetVault, PartialVault};
+pub use vault::{AssetVault, AssetWitness, PartialVault};
 
 // ASSET
 // ================================================================================================