+++ crates/miden-tx/src/executor/data_store.rs
@@ -1,6 +1,7 @@
 use alloc::collections::BTreeSet;
 
-use miden_objects::account::{Account, AccountId};
+use miden_objects::account::{AccountId, PartialAccount};
+use miden_objects::asset::AssetWitness;
 use miden_objects::block::{BlockHeader, BlockNumber};
 use miden_objects::transaction::PartialBlockchain;
 use miden_processor::{FutureMaybeSend, MastForestStore, Word};
@@ -31,6 +32,18 @@ pub trait DataStore: MastForestStore {
         account_id: AccountId,
         ref_blocks: BTreeSet<BlockNumber>,
     ) -> impl FutureMaybeSend<
-        Result<(Account, Option<Word>, BlockHeader, PartialBlockchain), DataStoreError>,
+        Result<(PartialAccount, Option<Word>, BlockHeader, PartialBlockchain), DataStoreError>,
     >;
+
+    /// Returns a witness for an asset in the requested account's vault with the requested vault
+    /// root.
+    ///
+    /// This is the witness that needs to be added to the advice provider's merkle store and advice
+    /// map to make access to the specified asset possible.
+    fn get_vault_asset_witness(
+        &self,
+        account_id: AccountId,
+        vault_root: Word,
+        vault_key: Word,
+    ) -> impl FutureMaybeSend<Result<AssetWitness, DataStoreError>>;
 }